requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fastapi==0.109.2
fonttools==4.49.0
h11==0.14.0
httptools==0.6.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
